Added conftest.py and unified labs fixtures into one place
