+++ tests/test_repdb_search.py
@@ -1,20 +1,5 @@
 import pytest
-import sys, os
-import json
-sys.path.insert(1, os.path.join(sys.path[0], '..'))
 
-from inquestlabs import inquestlabs_api
-from inquestlabs import inquestlabs_exception
-
-@pytest.fixture
-def labs():
-    labs = inquestlabs_api()
-    return labs
-
-@pytest.fixture
-def labs_with_key():
-    labs_api = inquestlabs_api(api_key="mock")
-    return labs_api
 
 @pytest.fixture
 def mock_list():
